Added new functionalities and minor improvements
Added an auto scan function. It runs the scan on the first ban on our
team automatically each game.

Added a replace image function. Usefull for fixing the database when adding the
wrong name picture for a player. It accepts player index and player
name, checks if the player exists and if it does updates its image with
the one with the given index.

Added a show player name image function. Usefull for checking if a given
player has the correct name image. Accepts a player name and shows a
window with the name image, which is closed on any button click.

Updated different pixels tollerance to 55 and required similarity index
to 0.7 to fix a bug where a player present in the data base was not
recognized.

Updated the sanitization of player 4 name since it was including a part
of the frame.

Updated logging. We now print "Scan" when running a scan. Changed the
delimiters of names and scans for easier read. The script now prints P0
heroes in red since these are the most important bans